Chp1 and chp2
Added references to chp.1 and started writing the SO/SE section in
chapter 2.
Cleaned up the text in chp.2 so that only relevant parts of summary are
left.
None of this will be used in actually delivery, since it needs to be
filtered even more and re-written (mainly just cleaned up the text
removing parts that aren't relevant at all)
each paragraph text has the source its taken from at the end to make it
faster for me to reference it later on