renamed chp5 to chp4, cleanup and re-writing chp4
Removed a lot of old comments to speed up reading process internally in
editor.
Deleted the old chapter4 (empty), and renamed chapter5 to chapter4, and
chapter6 to chapter5 (filename only). Also working on re-writing all of
chapter4, due to the messy structure.
The part about gridsearch and exhaustive search were moved to
methodology